backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.0.1
playwright==1.49.1
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.41.1
ultralytics
numpy
python-docx